requirements.txt
torch
fschat
pandas
seaborn
matplotlib
numpy
datasets
evaluate
anthropic
openai
python-dotenv
attr-dot-dict

transformers
google-generativeai
